moved stub handling to its own module
